refactor(chart-intelligence): extract the blend's source computation into TierListBlendBuilder
Pure extraction ahead of the Personalized Breakdown query: lens weight tables,
the K7 skill inference, the competitive-cohort similar players, and the
per-chart combine move to a shared internal builder that also surfaces the
intermediates (pooled skill deviations + evidence, source active flags,
neighbor counts) the breakdown needs. BlendedTierListHandler keeps lens
validation, the final combine, and the cache; its ctor and behavior are
unchanged — the existing handler tests pass untouched.
